+++ src/main.rs
@@ -1,16 +1,18 @@
 use std::env::args;
 use std::path::PathBuf;
 
+use mt63::wav;
+
 fn main() -> anyhow::Result<()> {
     let args = args().collect::<Vec<_>>();
     let wav_path = args.get(1).expect("Need a path to a wave file");
     let wav_path = wav_path.parse::<PathBuf>()?;
 
-    let (channel, sample_rate) = mt63::get_first_channel_from_wav(&wav_path)?;
+    let (channel, sample_rate) = wav::get_first_channel_from_wav(&wav_path)?;
     let mut offset = 1_000_000;
 
     let num_samples = 44;
-    let generated_samples = mt63::generate_samples(2500.0, sample_rate, num_samples);
+    let generated_samples = wav::generate_samples(2500.0, sample_rate, num_samples);
 
     for _ in 0..10 {
         let samples = channel
@@ -39,17 +41,4 @@ fn main() -> anyhow::Result<()> {
     }
     // dbg!(&sin_wave);
     Ok(())
-}
-
-fn augmented_hadamard_matrix() -> [u64; 128] {
-    let mut matrix: [u64; 128] = [0; 128];
-    for i in 0..matrix.len() {
-        for j in 0..64 {
-            let bits = i & j;
-            if (bits.count_ones() & 0x1 == 0) == (i & 0x40 == 0) {
-                matrix[i] = matrix[i] | (0x1 << (63 - j));
-            }
-        }
-    }
-    matrix
 }
